Rename VisibleRender to VisibilityObserver
