+++ src/components/VisibilityObserver/index.ts
@@ -1,3 +1,3 @@
 /*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
-export { default } from './VisibleRender';
-export * from './VisibleRender';
+export { default } from './VisibilityObserver';
+export * from './VisibilityObserver';
